Create Main.java
In order to make it a 60-70% win rate, I made the boundaries for intelligence from 10-52, because 19/30=63%, and 11/30=37%, therefore in order for the randomization process to have a 60%-70% chance of getting a 19 or above, the boundaries must be shifted to 52 (19/52 = 36%, and 33/52=63.5%). I then tested for the intelligence level to be over 30, and if it was, it would simply be changed back down to a level 30, as to not bypass the 50 total level maximum. This proved to be successful in making the randomized process a 60-70% win rate.